refactor(providers): Extract filters from reader
