+++ src/providers/reader.ts
@@ -2,20 +2,26 @@ import * as path from 'path';
 
 import micromatch = require('micromatch');
 
+import DeepFilter from './filters/deep';
+import EntryFilter from './filters/entry';
+
 import { IOptions } from '../managers/options';
 import { ITask } from '../managers/tasks';
 
-import { IEntry, IReaddirOptions } from 'readdir-enhanced';
+import { IReaddirOptions } from 'readdir-enhanced';
 import { Entry, EntryItem } from '../types/entries';
-import { Pattern } from '../types/patterns';
 
 export default abstract class Reader {
-	public readonly index: Map<string, undefined> = new Map();
-
 	private readonly micromatchOptions: micromatch.Options;
+
+	private readonly entryFilter: EntryFilter;
+	private readonly deepFilter: DeepFilter;
 
 	constructor(public readonly options: IOptions) {
 		this.micromatchOptions = this.getMicromatchOptions();
+
+		this.entryFilter = new EntryFilter(options, this.micromatchOptions);
+		this.deepFilter = new DeepFilter(options, this.micromatchOptions);
 	}
 
 	/**
@@ -36,8 +42,8 @@ export default abstract class Reader {
 	public getReaderOptions(task: ITask): IReaddirOptions {
 		return {
 			basePath: task.base === '.' ? '' : task.base,
-			filter: (entry) => this.filter(entry, task.patterns, task.negative),
-			deep: (entry) => this.deep(entry, task.negative, task.globstar),
+			filter: (entry) => this.entryFilter.call(entry, task.patterns, task.negative),
+			deep: (entry) => this.deepFilter.call(entry, task.negative, task.globstar),
 			sep: '/'
 		};
 	}
@@ -54,72 +60,6 @@ export default abstract class Reader {
 			nocase: this.options.nocase,
 			matchBase: this.options.matchBase
 		};
-	}
-
-	/**
-	 * Returns true if entry must be added to result.
-	 */
-	public filter(entry: IEntry, patterns: Pattern[], negative: Pattern[]): boolean {
-		// Exclude duplicate results
-		if (this.options.unique) {
-			if (this.isDuplicateEntry(entry)) {
-				return false;
-			}
-
-			this.createIndexRecord(entry);
-		}
-
-		// Mark directory by the final slash. Need to micromatch to support «directory/**» patterns
-		if (entry.isDirectory()) {
-			entry.path += '/';
-		}
-
-		// Filter directories that will be excluded by deep filter
-		if (entry.isDirectory() && micromatch.any(entry.path, negative, this.micromatchOptions)) {
-			return false;
-		}
-
-		// Filter files and directories by options
-		if (this.onlyFileFilter(entry) || this.onlyDirectoryFilter(entry)) {
-			return false;
-		}
-
-		// Filter by patterns
-		const entries = micromatch([entry.path], patterns, this.micromatchOptions);
-
-		return entries.length !== 0;
-	}
-
-	/**
-	 * Returns true if directory must be read.
-	 */
-	public deep(entry: IEntry, negative: Pattern[], globstar: boolean): boolean {
-		if (!this.options.deep) {
-			return false;
-		}
-
-		// Skip reading, depending on the nesting level
-		if (typeof this.options.deep === 'number') {
-			if (entry.depth > this.options.deep) {
-				return false;
-			}
-		}
-
-		// Skip reading if the directory is symlink and we don't want expand symlinks
-		if (this.isFollowedSymlink(entry)) {
-			return false;
-		}
-
-		// Skip reading if the directory name starting with a period and is not expected
-		if (this.isFollowedDotDirectory(entry)) {
-			return false;
-		}
-
-		if (micromatch.any(entry.path, negative, this.micromatchOptions)) {
-			return false;
-		}
-
-		return globstar;
 	}
 
 	/**
@@ -148,57 +88,5 @@ export default abstract class Reader {
 	 */
 	public isEnoentCodeError(err: NodeJS.ErrnoException): boolean {
 		return err.code === 'ENOENT';
-	}
-
-	/**
-	 * Returns true if the last partial of the path starting with a period.
-	 */
-	public isDotDirectory(entry: IEntry): boolean {
-		const pathPartials = entry.path.split('/');
-		const lastPathPartial: string = pathPartials[pathPartials.length - 1];
-
-		return lastPathPartial.startsWith('.');
-	}
-
-	/**
-	 * Return true if the entry already has in the cross reader index.
-	 */
-	public isDuplicateEntry(entry: IEntry): boolean {
-		return this.index.has(entry.path);
-	}
-
-	/**
-	 * Create record in the cross reader index.
-	 */
-	private createIndexRecord(entry: IEntry): void {
-		this.index.set(entry.path, undefined);
-	}
-
-	/**
-	 * Returns true for non-files if the «onlyFiles» option is enabled.
-	 */
-	private onlyFileFilter(entry: IEntry): boolean {
-		return this.options.onlyFiles && !entry.isFile();
-	}
-
-	/**
-	 * Returns true for non-directories if the «onlyDirectories» option is enabled.
-	 */
-	private onlyDirectoryFilter(entry: IEntry): boolean {
-		return this.options.onlyDirectories && !entry.isDirectory();
-	}
-
-	/**
-	 * Returns true for dot directories if the «dot» option is enabled.
-	 */
-	private isFollowedDotDirectory(entry: IEntry): boolean {
-		return !this.options.dot && this.isDotDirectory(entry);
-	}
-
-	/**
-	 * Returns true for symlinked directories if the «followSymlinks» option is enabled.
-	 */
-	private isFollowedSymlink(entry: IEntry): boolean {
-		return !this.options.followSymlinkedDirectories && entry.isSymbolicLink();
 	}
 }
